consolidate utils, enumerations, and context_processors
