+++ ohmg/content/models.py
@@ -30,7 +30,7 @@ from cogeo_mosaic.mosaic import MosaicJSON
 from cogeo_mosaic.backends import MosaicBackend
 
 from ohmg.georeference.models.resources import Layer
-from ohmg.georeference.utils import random_alnum
+from ohmg.utils import random_alnum
 from ohmg.loc_insurancemaps.models import Volume, Sheet
 
 logger = logging.getLogger(__name__)
