+++ ohmg/loc_insurancemaps/utils.py
@@ -2,7 +2,6 @@ import os
 import json
 import pytz
 import time
-import shutil
 import logging
 import requests
 from datetime import datetime
@@ -11,7 +10,12 @@ from PIL import Image
 from django.conf import settings
 from django.urls import reverse
 
-from .enumerations import (
+from ohmg.utils import (
+    download_image,
+    full_capitalize,
+)
+
+from ohmg.utils import (
     STATE_CHOICES,
 )
 
@@ -98,26 +102,6 @@ def unsanitize_name(state, name):
 
     return rev.get(name, name)
 
-def full_capitalize(in_str):
-    return " ".join([i.capitalize() for i in in_str.split(" ")])
-
-def download_image(url, out_path, retries=3):
-
-    # basic download code: https://stackoverflow.com/a/18043472/3873885
-    while True:
-        logger.debug(f"request {url}")
-        response = requests.get(url, stream=True)
-        if response.status_code == 200:
-            with open(out_path, 'wb') as out_file:
-                shutil.copyfileobj(response.raw, out_file)
-            return out_path
-        else:
-            logger.warn(f"response code: {response.status_code} retries left: {retries}")
-            time.sleep(5)
-            retries -= 1
-            if retries == 0:
-                logger.warn(f"request failed, cancelling")
-                return None
 
 class LOCParser(object):
