+++ ohmg/settings.py
@@ -111,8 +111,8 @@ TEMPLATES = [
         "django.contrib.auth.context_processors.auth",
         "django.contrib.messages.context_processors.messages",
         "django.contrib.auth.context_processors.auth",
-        "ohmg.loc_insurancemaps.context_processors.loc_info",
-        "ohmg.loc_insurancemaps.context_processors.general",
+        "ohmg.frontend.context_processors.loc_info",
+        "ohmg.frontend.context_processors.general",
         "pinax_theme_bootstrap.context_processors.theme",
       ],
       "debug": DEBUG,
